core_api/requirements.txt
fastapi
uvicorn
prisma
httpx
pydantic
passlib
bcrypt 
pyjwt
python-multipart